se agrego seguridad y funcianalidades al sistema de participaciones

+++ php/baja_participacion.php
@@ -1,4 +1,9 @@
 <?php
+$verificacion_participacion=mysqli_query($conexion,"SELECT *
+                        from participacion
+                        where idParticipacion='$_POST[idParticipacion]'")
+                        or die ("eeror en la verificacion");
+if  (mysqli_fetch_array($verificacion_participacion)['estado'] == 2){
   if ($_POST['estado'] == 1){
   mysqli_query($conexion,"DELETE from participacion where idParticipacion='$_POST[idParticipacion]'")
   or die ("error en la carga de participacion");
@@ -11,7 +16,7 @@
                             values (0,'$_SESSION[userId]',3,now(),-1,'penalizacion por cancelacion de participacion')")
                             or die ("error en penalizacion");
   }
-
+}
   setcookie("baja_participacion",true);
   $r = new Router;
   $file = $r->get_file();
